requirements.txt
torch
opencv-python
gym
gym-retro
stable-baselines3[extra]
